+++ video.h
@@ -24,6 +24,9 @@ BackgroundSubtractorKNN BackgroundSubtractorKNN_Create();
 void BackgroundSubtractorKNN_Close(BackgroundSubtractorKNN b);
 void BackgroundSubtractorKNN_Apply(BackgroundSubtractorKNN b, Mat src, Mat dst);
 
+void CalcOpticalFlowPyrLK(Mat prevImg, Mat nextImg, Mat prevPts, Mat nextPts, Mat status, Mat err);
+void CalcOpticalFlowFarneback(Mat prevImg, Mat nextImg, Mat flow, double pyrScale, int levels, int winsize,
+	int iterations, int polyN, double polySigma, int flags);
 #ifdef __cplusplus
 }
 #endif